requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
pydantic==2.6.4
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
openai
chromadb
python-multipart
pypdf
langchain==0.1.20
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
torch --index-url https://download.pytorch.org/whl/cpu
pypdf
python-dotenv